microservices/shared/requirements.txt
# Shared dependencies for Buffr Host microservices
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.0
fastapi-limiter>=0.1.5
